Add new header: D2Combat.h
