+++ source/D2Common/include/D2Combat.h
@@ -1,0 +1,32 @@
+#pragma once
+
+enum D2C_DamageReductionTypes
+{
+	DMGREDUCTYPE_NONE,
+	DMGREDUCTYPE_PHYSICAL,
+	DMGREDUCTYPE_MAGICAL,
+};
+
+enum D2C_DamageApplyTypes
+{
+	DMGAPPLYTYPE_DURATION,
+	DMGAPPLYTYPE_DIRECTDMG,
+};
+
+enum D2C_ElemTypes
+{
+	ELEMTYPE_NONE,
+	ELEMTYPE_FIRE,
+	ELEMTYPE_LTNG,
+	ELEMTYPE_MAGIC,
+	ELEMTYPE_COLD,
+	ELEMTYPE_POIS,
+	ELEMTYPE_LIFE,
+	ELEMTYPE_MANA,
+	ELEMTYPE_STAMINA,
+	ELEMTYPE_STUN,
+	ELEMTYPE_RAND,
+	ELEMTYPE_BURN,
+	ELEMTYPE_FREEZE,
+};
+
