+++ source/D2CommonDefinitions/include/D2Constants.h
@@ -2101,38 +2101,6 @@ enum D2Killtypes
 };
 #endif // CONSTANTS_ARENA
 
-#ifndef CONSTANTS_COMBAT //Pseudo-Macro to group all constants
-enum D2C_DamageReductionTypes
-{
-	DMGREDUCTYPE_NONE,
-	DMGREDUCTYPE_PHYSICAL,
-	DMGREDUCTYPE_MAGICAL,
-};
-
-enum D2C_DamageApplyTypes
-{
-	DMGAPPLYTYPE_DURATION,
-	DMGAPPLYTYPE_DIRECTDMG,
-};
-
-enum D2C_ElemTypes
-{
-	ELEMTYPE_NONE,
-	ELEMTYPE_FIRE,
-	ELEMTYPE_LTNG,
-	ELEMTYPE_MAGIC,
-	ELEMTYPE_COLD,
-	ELEMTYPE_POIS,
-	ELEMTYPE_LIFE,
-	ELEMTYPE_MANA,
-	ELEMTYPE_STAMINA,
-	ELEMTYPE_STUN,
-	ELEMTYPE_RAND,
-	ELEMTYPE_BURN,
-	ELEMTYPE_FREEZE,
-};
-#endif // CONSTANTS_COMBAT
-
 #ifndef CONSTANTS_PACKETS //Pseudo-Macro to group all constants
 enum D2C_SRV2CLT5A_TYPES
 {
